requirements.txt
streamlit
openai>=1.0.0
pandas
tiktoken
python-docx
numpy
PyPDF2
pytest
